Seperated the game into scenes
I made a GameSceneManager that will contain all the code related to
loading different scenes. You will notice that when loading a scenet
there are two Network modes, LOCAL and NETWORK. Currently network is not
implemented. We still need to inject GameSceneManager with photon to
implement that mode. However the LOCAL modes should be working.

There is currently a bug with playing the game a second time. It looks
like MenuManager Instance is null after the first game so no new menus
can be open. The entire first playthrough should work.